requirements.txt
django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
torch==2.0.0
drf-yasg==1.21.0
diffusers==0.25.0
djangorestframework-simplejwt==5.3.0
accelerate==0.27.0
transformers==4.36.0
sentencepiece==0.1.99
protobuf==4.25.1
sacremoses==0.0.53 